integrated next with the shop
