+++ package.json
@@ -4,14 +4,13 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"next": "latest",
-		"react": "latest",
-		"react-dom": "latest",
 		"@testing-library/jest-dom": "^5.16.5",
 		"@testing-library/react": "^13.4.0",
 		"@testing-library/user-event": "^13.5.0",
 		"gh-pages": "^5.0.0",
-		"react-router-dom": "^6.10.0",
+		"next": "^13.3.0",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
 		"react-scripts": "5.0.1",
 		"sass": "^1.61.0",
 		"web-vitals": "^2.1.4"
