+++ pages/FeaturesPage.js
@@ -1,0 +1,21 @@
+
+import VideoSection from '../src/components/VideoSeciton';
+import Products from '../src/components/Products';
+import Newsletter from '../src/components/Newsletter';
+import ProductsSoon from '../src/components/ProductsSoon';
+import Map from '../src/components/Map';
+
+
+const FeaturesPage = () => {
+   return (
+      <>
+         <VideoSection />
+         <Products />
+         <Newsletter />
+         <ProductsSoon />
+         <Map />
+      </>
+   );
+}
+
+export default FeaturesPage;
