+++ pages/PricingPage.js
