+++ pages/_app.js
@@ -1,0 +1,8 @@
+
+import Layout from "../src/components/Layout/Layout"
+
+function App({ Component, pageProps }) {
+  return <Layout><Component {...pageProps} /></Layout>
+}
+
+export default App
