+++ pages/index.js
@@ -1,0 +1,20 @@
+
+function HomePage() {
+  return (
+    <div className="app">
+      {/* <BrowserRouter> */}
+       
+<h1> hello our Shop App</h1>
+        {/* <Routes>
+          <Route path='/' element={<FeaturesPage />} />
+          <Route path='/pricing' element={<PricingPage />} />
+          <Route path='/community' element={<CommunityPage />} />
+          <Route path='/support' element={<SupportPage />} />
+        </Routes> */}
+      {/* </BrowserRouter> */}
+      {/* <Footer /> */}
+    </div>
+  );
+}
+
+export default HomePage;
