+++ public/assets/sample-logo.png
