--- src/components/FeaturesPage.js
@@ -1,21 +1,0 @@
-
-import VideoSection from './VideoSeciton';
-import Products from './Products';
-import Newsletter from './Newsletter';
-import ProductsSoon from './ProductsSoon';
-import Map from './Map';
-
-
-const FeaturesPage = () => {
-   return (
-      <>
-         <VideoSection />
-         <Products />
-         <Newsletter />
-         <ProductsSoon />
-         <Map />
-      </>
-   );
-}
-
-export default FeaturesPage;
