+++ src/components/Layout/Layout.js
@@ -1,0 +1,15 @@
+import NavBar from './NavBar'; 
+import Footer from './Footer'; 
+
+
+function Layout(props) {
+  return (
+    <div>
+      <NavBar />
+      <main >{props.children}</main>
+      <Footer />
+    </div>
+  );
+}
+
+export default Layout;
